Move MouseState to own file
