+++ src/Veldrid.SDL2/Sdl2Window.cs
@@ -1,15 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Numerics;
 using System.Runtime.CompilerServices;
-using System.Threading;
-using System.Threading.Tasks;
-using System.Diagnostics;
 using System.Text;
-
+using System.Threading;
 using static Veldrid.Sdl2.Sdl2Native;
-using System.ComponentModel;
-using Veldrid;
 
 namespace Veldrid.Sdl2
 {
@@ -37,6 +33,8 @@ namespace Veldrid.Sdl2
         {
         }
 
+        public Vector2 MouseDelta => _currentMouseDelta;
+
         public event Action<MouseWheelEventArgs> MouseWheel;
         public event Action<MouseMoveEventArgs> MouseMove;
         public event Action<MouseEvent> MouseDown;
@@ -54,8 +52,6 @@ namespace Veldrid.Sdl2
                 _currentMouseY = y;
             }
         }
-
-        public Vector2 MouseDelta => _currentMouseDelta;
 
         public InputSnapshot PumpEvents()
         {
@@ -250,106 +246,6 @@ namespace Veldrid.Sdl2
                 other.WheelDelta = WheelDelta;
                 _mouseDown.CopyTo(other._mouseDown, 0);
             }
-        }
-    }
-
-    public struct MouseState
-    {
-        public readonly int X;
-        public readonly int Y;
-
-        private bool _mouseDown0;
-        private bool _mouseDown1;
-        private bool _mouseDown2;
-        private bool _mouseDown3;
-        private bool _mouseDown4;
-        private bool _mouseDown5;
-        private bool _mouseDown6;
-        private bool _mouseDown7;
-        private bool _mouseDown8;
-        private bool _mouseDown9;
-        private bool _mouseDown10;
-        private bool _mouseDown11;
-        private bool _mouseDown12;
-
-        public MouseState(
-            int x, int y,
-            bool mouse0, bool mouse1, bool mouse2, bool mouse3, bool mouse4, bool mouse5, bool mouse6,
-            bool mouse7, bool mouse8, bool mouse9, bool mouse10, bool mouse11, bool mouse12)
-        {
-            X = x;
-            Y = y;
-            _mouseDown0 = mouse0;
-            _mouseDown1 = mouse1;
-            _mouseDown2 = mouse2;
-            _mouseDown3 = mouse3;
-            _mouseDown4 = mouse4;
-            _mouseDown5 = mouse5;
-            _mouseDown6 = mouse6;
-            _mouseDown7 = mouse7;
-            _mouseDown8 = mouse8;
-            _mouseDown9 = mouse9;
-            _mouseDown10 = mouse10;
-            _mouseDown11 = mouse11;
-            _mouseDown12 = mouse12;
-        }
-
-        public bool IsButtonDown(MouseButton button)
-        {
-            uint index = (uint)button;
-            switch (index)
-            {
-                case 0:
-                    return _mouseDown0;
-                case 1:
-                    return _mouseDown1;
-                case 2:
-                    return _mouseDown2;
-                case 3:
-                    return _mouseDown3;
-                case 4:
-                    return _mouseDown4;
-                case 5:
-                    return _mouseDown5;
-                case 6:
-                    return _mouseDown6;
-                case 7:
-                    return _mouseDown7;
-                case 8:
-                    return _mouseDown8;
-                case 9:
-                    return _mouseDown9;
-                case 10:
-                    return _mouseDown10;
-                case 11:
-                    return _mouseDown11;
-                case 12:
-                    return _mouseDown12;
-            }
-
-            throw new ArgumentOutOfRangeException(nameof(button));
-        }
-    }
-
-    public struct MouseWheelEventArgs
-    {
-        public MouseState State { get; }
-        public float WheelDelta { get; }
-        public MouseWheelEventArgs(MouseState mouseState, float wheelDelta)
-        {
-            State = mouseState;
-            WheelDelta = wheelDelta;
-        }
-    }
-
-    public struct MouseMoveEventArgs
-    {
-        public MouseState State { get; }
-        public Vector2 MousePosition { get; }
-        public MouseMoveEventArgs(MouseState mouseState, Vector2 mousePosition)
-        {
-            State = mouseState;
-            MousePosition = mousePosition;
         }
     }
 }
